requirements.txt
# FastAPI for the web framework
fastapi==0.95.0
uvicorn==0.23.2
pydantic==1.10.7
python-dotenv==1.0.0
httpx==0.24.1

# Google Calendar API
google-api-python-client==2.107.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Testing
pytest==7.4.3

# Optional - for better logging
loguru==0.7.2